+++ src/components/Table/Table.stories.tsx
@@ -35,6 +35,39 @@ const fakeData = [
   },
 ];
 
+const items = [
+  {
+    label: 'Angry Lama',
+    onClick: () => {
+      console.log('Item 1 clicked');
+    },
+  },
+  {
+    label: 'Frightened Frog',
+    onClick: () => {
+      console.log('Item 2 clicked');
+    },
+  },
+  {
+    label: 'Hungry Bear',
+    onClick: () => {
+      console.log('Item 3 clicked');
+    },
+  },
+  {
+    label: 'Crazy Koala',
+    onClick: () => {
+      console.log('Item 4 clicked');
+    },
+  },
+  {
+    label: 'Dizzy Snake',
+    onClick: () => {
+      console.log('Item 5 clicked');
+    },
+  },
+];
+
 export const Standard = () => (
   <Table>
     <Table.Row>
@@ -61,7 +94,10 @@ export const WithActions = () => (
         <Table.Cell>{data.count}</Table.Cell>
         <Table.Cell>{data.name}</Table.Cell>
         <Table.Cell>
-          <Table.Action onClick={() => {}} text='filter' />
+          <Table.ActionContainer>
+            <Table.Action onClick={() => {}} text='filter' />
+            <Table.ActionMenu items={items} />
+          </Table.ActionContainer>
         </Table.Cell>
       </Table.Row>
     ))}