+++ src/components/Table/_TableActionContainer.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const TableActionContainer = styled.div`
+  display: flex;
+`;
+
+export default TableActionContainer;